+++ src/engine/DiffAlgorithms.cpp
@@ -20,7 +20,7 @@ int DiffAlgorithms::GetInstructionHashMatchRate(vector<unsigned char> instructio
 {
 	int matchRate = 0;
 
-	int lengthDifference = (instructionHash1.size() - instructionHash2.size());
+	size_t lengthDifference = (instructionHash1.size() - instructionHash2.size());
 	if (lengthDifference > instructionHash1.size() * 0.5 || lengthDifference > instructionHash2.size() * 0.5)
 	{
 		matchRate = 0;
